Update round buttons
Include more if/else checks

+++ app/include/r1_end.php
@@ -1,17 +1,28 @@
 <?php 
     require_once 'common.php';
     require_once 'clearing1.php';
-
-    first_clearing();
     $roundDAO = new RoundDAO();
-    $bidDAO = new BidDAO;
-    $bidDAO->removeAll();
-    $round = 1;
-    $status = "inactive";
+    $round = $roundDAO->retrieveAll();
 
-    $update_status = $roundDAO->updateStatus($round,$status);
-    if ($update_status){
-        echo "Round 1 has ended!";
+    if ( ($round->round == 1) && ($round->status == 'active') ){
+        first_clearing();
+        $roundDAO = new RoundDAO();
+        $bidDAO = new BidDAO();
+        $bidDAO->removeAll();
+        $round = 1;
+        $status = "inactive";
+
+        
+        $update_status = $roundDAO->updateStatus($round,$status);
+        if ($update_status){
+            echo "Round 1 has ended!";
+        }
+        echo "<a href = '../admin_homepage.php'> Back </a>";
     }
-    echo "<a href = '../admin_homepage.php'> Back </a>";
+    else{
+        echo "round already ended";
+        echo "<a href = '../admin_homepage.php'> Back </a>";
+    }
+
+    
 ?>
