+++ app/include/r2_end.php
@@ -2,16 +2,26 @@
     require_once 'common.php';
     require_once 'clearing1.php';
 
-    first_clearing();
     $roundDAO = new RoundDAO();
-    $bidDAO = new BidDAO;
-    $bidDAO->removeAll();
-    $round = 2;
-    $status = "inactive";
+    $round = $roundDAO->retrieveAll();
 
-    $update_status = $roundDAO->updateStatus($round,$status);
-    if ($update_status){
-        echo "Round 2 has ended!";
+    if ( ($round->round == 2) && ($round->status == 'active')){
+        first_clearing();
+        $roundDAO = new RoundDAO();
+        $bidDAO = new BidDAO;
+        $bidDAO->removeAll();
+        $round = 2;
+        $status = "inactive";
+
+        $update_status = $roundDAO->updateStatus($round,$status);
+        if ($update_status){
+            echo "Round 2 has ended!";
+        }
+        echo "<a href = '../admin_homepage.php'> Back </a>";
     }
-    echo "<a href = '../admin_homepage.php'> Back </a>";
+    else{
+        echo "round already ended";
+        echo "<a href = '../admin_homepage.php'> Back </a>";
+    }
+    
 ?>
